+++ NEWS.md
@@ -1,7 +1,7 @@
 # protti ***
 
 ## Bug fixes
-* `try_query()` now checks if a valid URL is provided before attempting to download data. 
+* `try_query()` now correctly handles errors that don't return a response object. We also handle gzip decompression problems better since some databases compressed responses were not handled correctly. 
 
 # protti 0.9.0