+++ NEWS.md
@@ -1,4 +1,4 @@
-# protti ***
+# protti 0.9.1
 
 ## Bug fixes
 * `try_query()` now correctly handles errors that don't return a response object. We also handle gzip decompression problems better since some databases compressed responses were not handled correctly.